#1824: [Fix] REST - Session timeout not working

+++ services/rest/v2/src/main/java/org/cmdbuild/service/rest/v2/cxf/service/InMemorySessionStore.java
@@ -18,7 +18,7 @@ public class InMemorySessionStore implements SessionStore {
 
 	public static interface Configuration {
 
-		int timeout();
+		long timeout();
 
 	}
